remove warning on g++

+++ include/Debug.h
@@ -16,7 +16,7 @@ void dumpInstVal(unsigned Val) {
 
   std::cerr << "\n";
   std::cerr << "Hex(LE): ";
-  for (int i = 0; i < sizeof(Val); ++i) {
+  for (unsigned long i = 0; i < sizeof(Val); ++i) {
     unsigned char byte = (Val >> (i * 8)) & 0xFF;
     std::cerr << std::hex << std::setw(2) << std::setfill('0')
               << static_cast<int>(byte) << ' ';
@@ -46,7 +46,7 @@ void debugInstOnAsm(const std::vector<std::string> Toks, unsigned Val) {
 
   std::cerr << "\n";
   std::cerr << "Hex(LE): ";
-  for (int i = 0; i < sizeof(Val); ++i) {
+  for (unsigned long i = 0; i < sizeof(Val); ++i) {
     unsigned char byte = (Val >> (i * 8)) & 0xFF;
     std::cerr << std::hex << std::setw(2) << std::setfill('0')
               << static_cast<int>(byte) << ' ';
